+++ CHANGELOG.md
@@ -2,10 +2,10 @@
 
 ## Unreleased
 
-### Added
+### Added
 - You can now specify the target section by it's UID or handle in addition to using ID.
 
-### Changed
+### Changed
 - Guest Entries now requires Craft 3.1.0-alpha.1 or later.
 
 ### Fixed