+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## Unreleased
 
 ### Added
-- You can now specify the target section by it's UID or handle in addition to using ID.
+- You can now specify the target section by its UID or handle in addition to using ID.
 
 ### Changed
 - Guest Entries now requires Craft 3.1.0-alpha.1 or later.