+++ CHANGELOG.md
@@ -2,11 +2,9 @@
 
 ## Unreleased
 
-### Added
-- You can now specify the target section by its UID or handle in addition to using ID.
-
 ### Changed
 - Guest Entries now requires Craft 3.1.0-alpha.1 or later.
+- You can now specify the target section by its UID or handle in addition to using ID.
 
 ### Fixed
 - Fixed a bug where it wasn’t possible to turn off guest submissions for a section once it had been activated.